refactor: move `prepareInclude` to AppController base class
