+++ plugins/BEdita/API/src/Controller/ResourcesController.php
@@ -25,7 +25,6 @@ use BEdita\Core\Model\Action\SetAssociatedAction;
 use BEdita\Core\Utility\JsonApiSerializable;
 use Cake\Core\InstanceConfigTrait;
 use Cake\Datasource\EntityInterface;
-use Cake\Network\Exception\BadRequestException;
 use Cake\Network\Exception\ConflictException;
 use Cake\Network\Exception\InternalErrorException;
 use Cake\Network\Exception\NotFoundException;
@@ -108,45 +107,6 @@ abstract class ResourcesController extends AppController
         }
 
         throw new NotFoundException(__d('bedita', 'Relationship "{0}" does not exist', $relationship));
-    }
-
-    /**
-     * Prepare a list of associations to be contained from `?include` query parameter.
-     *
-     * @param string|array|null $include Association(s) to be included.
-     * @return array
-     * @throws \Cake\Network\Exception\BadRequestException Throws an exception if a
-     */
-    protected function prepareInclude($include)
-    {
-        if ($include === null) {
-            return [];
-        }
-        if (!is_string($include)) {
-            throw new BadRequestException(
-                __d('bedita', 'Invalid "{0}" query parameter ({1})', 'include', __d('bedita', 'Must be a comma-separated string'))
-            );
-        }
-
-        $contain = [];
-        $include = array_filter(array_map('trim', explode(',', $include)));
-        foreach ($include as $relationship) {
-            if (strpos($relationship, '.') !== false) {
-                throw new BadRequestException(__d('bedita', 'Inclusion of nested resources is not yet supported'));
-            }
-
-            try {
-                $association = $this->findAssociation($relationship);
-            } catch (NotFoundException $e) {
-                throw new BadRequestException(
-                    __d('bedita', 'Invalid "{0}" query parameter ({1})', 'include', __d('bedita', 'Relationship "{0}" does not exist', $relationship))
-                );
-            }
-
-            $contain[] = $association->getName();
-        }
-
-        return $contain;
     }
 
     /**
